Перенос состояния загрузки данных из data в application
