+++ src/reducers/application/reducer.js
@@ -5,11 +5,13 @@ import {SortTypes} from '../../const.js';
 const initialState = {
   city: CITIES[0],
   currentSort: SortTypes.POPULAR,
+  isLoading: false,
 };
 
 const ActionType = {
   CHANGE_CITIES: `CHANGE_CITIES`,
   SORT_TYPE_CHANGE: `SORT_TYPE_CHANGE`,
+  CHANGE_STATUS_LOADING: `CHANGE_STATUS_LOADING`,
 };
 
 const ActionCreator = {
@@ -33,6 +35,10 @@ const ActionCreator = {
 
     return {};
   },
+  activatedPreloader: (isLoading) => ({
+    type: ActionType.CHANGE_STATUS_LOADING,
+    payload: !isLoading,
+  }),
 };
 
 const reducer = (store = initialState, action) => {
@@ -41,9 +47,15 @@ const reducer = (store = initialState, action) => {
       return extend(store, {
         city: action.payload
       });
+
     case ActionType.SORT_TYPE_CHANGE:
       return extend(store, {
         currentSort: action.payload
+      });
+
+    case ActionType.CHANGE_STATUS_LOADING:
+      return extend(store, {
+        isLoading: action.payload
       });
   }
