+++ src/reducers/application/selectors.js
@@ -1,17 +1,17 @@
 import NameSpace from '../name-space.js';
 
-const NAME_SPACE = NameSpace.OFFER;
+const NAME_SPACE = NameSpace.APPLICATION;
 
 const getCurrentCity = (state) => {
   return state[NAME_SPACE].city;
+};
+
+const getStatusLoading = (state) => {
+  return state[NAME_SPACE].isLoading;
 };
 
 const getCurrentSort = (state) => {
   return state[NAME_SPACE].currentSort;
 };
 
-const getLoadingStatus = (state) => {
-  return state[NAME_SPACE].isLoading;
-}
-
-export {getCurrentCity, getCurrentSort, getLoadingStatus};
+export {getCurrentCity, getCurrentSort, getStatusLoading};
