+++ src/reducers/data/reducer.js
@@ -1,14 +1,13 @@
 import {extend} from '../../utils/util.js';
 import {hotelsAdapter} from '../../adapters/adapters.js';
+import {ActionCreator as ActionCreatorApplication} from '../application/reducer.js';
 
 const initialState = {
   offers: [],
-  isLoading: false,
 };
 
 const ActionType = {
   GET_OFFERS: `GET_OFFERS`,
-  CHANGE_STATUS_LOADING: `CHANGE_STATUS_LOADING`,
 };
 
 const ActionCreator = {
@@ -16,20 +15,14 @@ const ActionCreator = {
     type: ActionType.GET_OFFERS,
     payload: offers
   }),
-
-  activatedPreloader: (isLoading) => ({
-    type: ActionType.CHANGE_STATUS_LOADING,
-    payload: !isLoading,
-  }),
-
 };
 
 const Operations = {
   loadOffers: () => (dispatch, getState, api) => {
     return api.get(`/hotels`)
       .then((response) => {
-        const isLoading = getState().DATA.isLoading;
-        dispatch(ActionCreator.activatedPreloader(isLoading));
+        const isLoading = getState().APPLICATION.isLoading;
+        dispatch(ActionCreatorApplication.activatedPreloader(isLoading));
 
         return response;
       })
@@ -46,10 +39,6 @@ const reducer = (store = initialState, action) => {
     case ActionType.GET_OFFERS:
       return extend(store, {
         offers: action.payload
-      });
-    case ActionType.CHANGE_STATUS_LOADING:
-      return extend(store, {
-        isLoading: action.payload
       });
   }
