+++ src/reducers/data/selectors.js
@@ -9,10 +9,6 @@ const getOffers = (state) => {
   return state[NAME_SPACE].offers;
 };
 
-const getStatusLoading = (state) => {
-  return state[NAME_SPACE].isLoading;
-};
-
 const getOffersAfterFiltration = createSelector(
   getOffers,
   getCurrentCity,
@@ -52,4 +48,4 @@ const getSortedOffers = createSelector(
     }
 );
 
-export {getSortedOffers, getStatusLoading};
+export {getSortedOffers};
